+++ src/Neo.SmartContract.Framework/Native/CryptoLib.cs
@@ -1,6 +1,6 @@
 #pragma warning disable CS0626
 
-namespace Neo.SmartContract.Framework.Services.Neo
+namespace Neo.SmartContract.Framework.Native
 {
     [Contract("0x726cb6e0cd8628a1350a611384688911ab75f51b")]
     public static class CryptoLib
