+++ src/Neo.SmartContract.Framework/Native/Ledger.cs
@@ -1,6 +1,8 @@
 #pragma warning disable CS0626
 
-namespace Neo.SmartContract.Framework.Services.Neo
+using Neo.SmartContract.Framework.Services;
+
+namespace Neo.SmartContract.Framework.Native
 {
     [Contract("0xda65b600f7124ce6c79950c1772a36403104f2be")]
     public class Ledger
