+++ src/Neo.SmartContract.Framework/Native/NamedCurve.cs
@@ -1,4 +1,4 @@
-namespace Neo.SmartContract.Framework.Services.Neo
+namespace Neo.SmartContract.Framework.Native
 {
     /// <summary>
     /// RFC 4492
