+++ src/Neo.SmartContract.Framework/Native/Oracle.cs
@@ -1,6 +1,6 @@
 #pragma warning disable CS0626
 
-namespace Neo.SmartContract.Framework.Services.Neo
+namespace Neo.SmartContract.Framework.Native
 {
     [Contract("0xfe924b7cfe89ddd271abaf7210a80a7e11178758")]
     public class Oracle
@@ -11,4 +11,3 @@ namespace Neo.SmartContract.Framework.Services.Neo
         public static extern void Request(string url, string filter, string callback, object userData, long gasForResponse);
     }
 }
-
