+++ src/Neo.SmartContract.Framework/Native/Policy.cs
@@ -1,6 +1,6 @@
 #pragma warning disable CS0626
 
-namespace Neo.SmartContract.Framework.Services.Neo
+namespace Neo.SmartContract.Framework.Native
 {
     [Contract("0xcc5e4edd9f5f8dba8bb65734541df7a1c081c67b")]
     public class Policy
