+++ src/Neo.SmartContract.Framework/Native/RecordType.cs
@@ -1,4 +1,4 @@
-namespace Neo.SmartContract.Framework.Services.Neo
+namespace Neo.SmartContract.Framework.Native
 {
     public enum RecordType : byte
     {
