+++ src/Neo.SmartContract.Framework/Native/RoleManagement.cs
@@ -1,6 +1,6 @@
 #pragma warning disable CS0626
 
-namespace Neo.SmartContract.Framework.Services.Neo
+namespace Neo.SmartContract.Framework.Native
 {
     [Contract("0x49cf4e5378ffcd4dec034fd98a174c5491e395e2")]
     public class RoleManagement
