+++ src/Neo.SmartContract.Framework/Services/Iterator.cs
@@ -1,4 +1,4 @@
-namespace Neo.SmartContract.Framework.Services.Neo
+namespace Neo.SmartContract.Framework.Services
 {
     public class Iterator
     {
