--- src/Neo.SmartContract.Framework/Services/Neo/TriggerType.cs
@@ -1,10 +1,0 @@
-namespace Neo.SmartContract.Framework.Services.Neo
-{
-    public enum TriggerType : byte
-    {
-        System = 0x01,
-        Verification = 0x20,
-        Application = 0x40,
-        All = System | Verification | Application
-    }
-}
