+++ src/Neo.SmartContract.Framework/Services/Storage.cs
@@ -1,6 +1,6 @@
 using System.Numerics;
 
-namespace Neo.SmartContract.Framework.Services.Neo
+namespace Neo.SmartContract.Framework.Services
 {
     public static class Storage
     {
