+++ src/Neo.SmartContract.Framework/SmartContract.cs
@@ -1,3 +1,4 @@
+using Neo.SmartContract.Framework.Native;
 using System.Runtime.CompilerServices;
 
 namespace Neo.SmartContract.Framework
@@ -18,5 +19,15 @@ namespace Neo.SmartContract.Framework
         [MethodImpl(MethodImplOptions.InternalCall)]
         [OpCode(OpCode.ABORT)]
         public static extern void Abort();
+
+        public static ByteString Hash160(ByteString value)
+        {
+            return CryptoLib.ripemd160(CryptoLib.Sha256(value));
+        }
+
+        public static ByteString Hash256(ByteString value)
+        {
+            return CryptoLib.Sha256(CryptoLib.Sha256(value));
+        }
     }
 }
